fix intermittent tests
