+++ indy_node/test/anon_creds/conftest.py
@@ -7,13 +7,12 @@ from indy.ledger import build_cred_def_request
 
 from indy_common.state.domain import make_state_path_for_claim_def
 from indy_node.test.anon_creds.helper import get_cred_def_id, create_revoc_reg, create_revoc_reg_entry
-from plenum.common.constants import DOMAIN_LEDGER_ID
 from plenum.common.util import randomString
 from indy_common.constants import REVOC_REG_ENTRY, REVOC_REG_DEF_ID, ISSUED, \
     REVOKED, PREV_ACCUM, ACCUM, REVOC_REG_DEF, ISSUANCE_BY_DEFAULT, \
-    CRED_DEF_ID, VALUE, TAG, ISSUANCE_ON_DEMAND, CLAIM_DEF, ID, \
+    CRED_DEF_ID, VALUE, TAG, ISSUANCE_ON_DEMAND, ID, \
     TXN_TYPE, REVOC_TYPE, ISSUANCE_TYPE, MAX_CRED_NUM, TAILS_HASH, TAILS_LOCATION, PUBLIC_KEYS, \
-    GET_REVOC_REG_DELTA, GET_REVOC_REG, TIMESTAMP, FROM, TO, CLAIM_DEF_SCHEMA_REF, CLAIM_DEF_SIGNATURE_TYPE, \
+    GET_REVOC_REG, TIMESTAMP, CLAIM_DEF_SCHEMA_REF, CLAIM_DEF_SIGNATURE_TYPE, \
     CLAIM_DEF_TAG
 from indy_common.types import Request
 from indy_common.state import domain
@@ -371,26 +370,4 @@ def send_revoc_reg_entry_by_demand(looper,
     return reg_entry_replies[0]
 
 
-@pytest.fixture(scope="module")
-def build_get_revoc_reg_entry(looper,
-                              sdk_wallet_steward):
-    data = {
-        REVOC_REG_DEF_ID: randomString(10),
-        TXN_TYPE: GET_REVOC_REG,
-        TIMESTAMP: int(time.time())
-    }
-    revoc_reg_req = sdk_sign_request_from_dict(looper, sdk_wallet_steward, data)
-    return revoc_reg_req
 
-
-@pytest.fixture(scope="module")
-def build_get_revoc_reg_delta(looper,
-                              sdk_wallet_steward):
-    data = {
-        REVOC_REG_DEF_ID: randomString(10),
-        TXN_TYPE: GET_REVOC_REG_DELTA,
-        FROM: 10,
-        TO: 20,
-    }
-    revoc_reg_delta_req = sdk_sign_request_from_dict(looper, sdk_wallet_steward, data)
-    return revoc_reg_delta_req
